outsource inputrender to a component
